==> code/02_list_ciks.ipynb <==
"""
all companies in S&P500 in the last 10 years
"""

import pandas as pd
import pickle

sp500 = pd.read_excel('../data/sp500_10yr.xlsx')
sp500.head()

# SEC list of tickers matched to cik's 
ciks = pd.read_csv("https://www.sec.gov/include/ticker.txt", header=None, 
                   delimiter='\t', names=['ticker', 'cik'], dtype={'cik':str})


ciks.head()

ciks["ticker"] = ciks.ticker.str.upper()
sp_500_ciks = pd.merge(sp500, ciks, how='left', left_on='Symbol', right_on='ticker')
sp_500_ciks.head()

with open('../data/sp_500_ciks.pickle', 'wb') as to_write:
    pickle.dump(sp_500_ciks, to_write)
