==> code/04_all_doc_links.ipynb <==
"""
combine the multiple dataframes created after SEC API data pull
"""

import pickle
import pandas as pd
import numpy as np

import os
import glob

def url_list_df(year):

    urls_ = pd.DataFrame()

    for file_name in glob.glob('../data/' + str(year) + '/year_*.pickle'):
        try:
            x = pd.read_pickle(file_name)
            urls_ = pd.concat([urls_,x],axis=0)
            print('APPENDED: ', file_name)
        except:
            print('ERROR: ', file_name)
            

    urls_.to_pickle('../data/urls_' + str(year) + '.pickle')
    print(year, ' files saved to pickle')
    
    urls_.to_csv('../data/urls_' + str(year) + '.csv', index=False)
    print(year, ' files saved to csv')

url_list_df(2020)

url_list_df(2019)

url_list_df(2018)

url_list_df(2017)

url_list_df(2016)

url_list_df(2015)

url_list_df(2014)

url_list_df(2013)

url_list_df(2012)

url_list_df(2011)
