==> code/05_10K_content_data_pull.ipynb <==
"""
pull data in the 10-K reports around the word 'risk'
ran for one year at a time 2011-2020
"""

import pickle
import pandas as pd
import numpy as np
from collections import defaultdict


from bs4 import BeautifulSoup
import re
import requests
import unicodedata
import random
import time

df = pd.read_pickle("../data/urls_2013.pickle")

df.head()

#create dict for webscrape below
form_dict = defaultdict(list)
for inx, row in df.iterrows():
    form_dict[row['ticker']].append(row['linkToTxt'])


text = {}

for ticker, urls in form_dict.items():
    for url in urls:
        res = requests.get(url)
        page = res.content
        
        try: 
            soup = BeautifulSoup(page, 'html.parser')
        except:
            text[ticker] = "error"
            continue

        # some documents use HTML div tags and others paragraph tags
        paragraphs = [paragraph for paragraph in soup.find_all('p')]
        divs = [div for div in soup.find_all('div')]

        blobs = []

        # search for "risk" or "risks"
        for paragraph in paragraphs:
            if paragraph.find_all(text=re.compile(r"(\brisk\b|\brisks\b)", re.IGNORECASE)): 
                for element in paragraph.find_all(text=re.compile(r"(\brisk\b|\brisks\b)", re.IGNORECASE)):
                    blobs.append(str(unicodedata.normalize('NFKD', element)))

        for div in divs:
            if div.find_all(text=re.compile(r"(\brisk\b|\brisks\b)", re.IGNORECASE)): 
                for element in div.find_all(text=re.compile(r"(\brisk\b|\brisks\b)", re.IGNORECASE)):
                    blobs.append(str(unicodedata.normalize('NFKD', element)))

        # remove any unparsed HTML 
        for blob in blobs: 
            if "font-family" in blob:
                blobs.remove(blob)

        text[ticker + ";" + url] = blobs 

        # pause before querying next document
        time.sleep(.5+2*random.random()) 
     
    
print(len(text))

with open('text_risk_data_2013.pickle', 'wb') as to_write:
    pickle.dump(text, to_write)
