==> code/06_preprocess_text.ipynb <==
"""
preprocess text extracted from 10K reports
ran one year at a time 2011-2020 to check output at each step
"""

import pandas as pd
import numpy as np
import pickle
import datetime

import re
import string

import nltk
from nltk.tokenize import word_tokenize
from nltk.tag import pos_tag

year = 2011

with open('text_risk_data_' + str(year) + '.pickle', 'rb') as read_file:
    text_dict = pickle.load(read_file)
    

#check for errors and remove
issues = {}
for key, value in text_dict.items():
    if value == "error":
        issues[key] = value

for key in issues.keys():
    del text_dict[key]

#into df
text_df = pd.DataFrame(dict([(k,pd.Series(v)) for k,v in text_dict.items()])).melt().dropna()

text_df[["ticker", "linkToTxt"]] = text_df["variable"].str.split(';', expand=True)
del text_df["variable"]

text_df.head()

# Remove duplicates
text_df2 = text_df.drop_duplicates(subset=None, keep='first', inplace=False, ignore_index=False)
text_df2.shape

#### CLEAN TEXT

# first couple years of data pulls didn't filter to word boundaries
text_df2.loc[:, 'keep'] = np.where(text_df2.loc[:, 'value'].str.contains( r"(\brisk\b|\brisks\b)", case=False, na=False), 1, 0)

text_df2.keep.value_counts()

the_keeps = text_df2[text_df2.keep == 1]
the_keeps.shape

# Text preprocessing steps - remove HTML tags, numbers, and punctuation

TAG_RE = re.compile(r'<[^>]+>')
def remove_tags(text):
    return TAG_RE.sub('', text)

the_keeps.loc[:,'cleaned' ] = the_keeps.loc[:,'value'].apply(remove_tags)

alphanumeric = lambda x: re.sub('\w*\d\w*', ' ', x)
punc_remove = lambda x: re.sub('[%s]' % re.escape(string.punctuation), ' ', x)

the_keeps.loc[:,'preproc_text'] = the_keeps.loc[:,'cleaned'].map(alphanumeric).map(punc_remove)


the_keeps.head()

# add text length and will check for very long strings, each df row should be ~ 1-2 sentences
the_keeps.loc[:,'value_len'] = the_keeps.loc[:,'value'].apply(lambda x : len(x))


# check for any very short text strings
the_keeps[the_keeps.value_len < 100].value.value_counts()

## some of the shorter strings are the section headers
## will remove the section headers
      # Risk Factors 
      # Quantitative and Qualitative Disclosures About Market Risk  
      # Quantitative and Qualitative Disclosures About Market Risks 
      # ITEM 1A. RISK FACTORS
      # ITEM 7A. QUANTITATIVE AND QUALITATIVE DISCLOSURES ABOUT MARKET RISK
        
the_keeps['value2'] = the_keeps['value'].apply(lambda x: " ".join(x.split()))

headers = ['risk factors', 'quantitative and qualitative disclosures about market risk',
          'quantitative and qualitative disclosures about market risk',
          'item 1a. risk factors', 'item 7a. quantitative and qualitative disclosures about market risk']    


f1_df = (the_keeps[~the_keeps['value2'].str.lower().isin(headers)])
f1_df.shape

# check for any very long text strings
f1_df[f1_df.value_len > 10_000].head()


len(f1_df[f1_df.value_len > 10_000].head())

# remove the strings over 10K that are some code and not text content
f2_df = f1_df[f1_df.value_len < 10_000]
f2_df.shape

#f2_df.to_excel('f2_df_review.xlsx', index=False)

################


f2_df.head()

f3_df = f2_df.drop(['value_len', 'value2', 'cleaned', 'keep'], axis=1, inplace=False)

f3_df.head()

with open('../data/preproc/preproc_text_' + str(year) + '.pickle', 'wb') as to_write:
    pickle.dump(f3_df, to_write)
