==> code/07_more_text_preprocessing.ipynb <==
''' 
for each year - merge df with text content with company info

stack all years

remove reports from companies from times when they were not in S&P 500 

lemmatize text

'''

import pickle
import pandas as pd
import numpy as np
import datetime
from nltk.stem import WordNetLemmatizer

all_text_data = pd.DataFrame()

for year in [2020, 2019, 2018, 2017, 2016, 2015, 2014, 2013, 2012, 2011]:
    
    # df of paragraphs about risk
    with open('../data/preproc/preproc_text_' + str(year) + '.pickle', 'rb') as read_file:
        prepped_text = pickle.load(read_file)
      
    # list with company info
    sp500_list = pd.read_pickle('../data/urls_' + str(year) + '.pickle')
    
    # merge 
    text_df_sp = pd.merge(prepped_text, sp500_list, how="left", on=["ticker", "linkToTxt"])
        
    #stack     
    all_text_data = pd.concat([all_text_data, text_df_sp], axis=0)

print(all_text_data.shape)
    

all_text_data.head()

# drop rows for companies that were not in S&P at time of annual report

# remove 3 added in 2021
f1_all = all_text_data[~all_text_data.ticker.isin(['MPWR', 'TRMB', 'ENPH'])]
f1_all.shape

# list of added tickers 
sp_adds = pd.read_excel('../sp500_changes.xlsx', sheet_name='added')
sp_adds.head()

sp_adds.info()

# check for duplicates
sp_adds.Added_Ticker.duplicated().value_counts()

# list of removed tickers
sp_removed = pd.read_excel('../sp500_changes.xlsx', sheet_name='removed')
sp_removed.head()

# check for duplicates
sp_removed.Removed_Ticker.duplicated().value_counts()

# merge add and removed dates with df
# flag rows to drop

print(f1_all.shape)

merge1 = pd.merge(f1_all, sp_adds, how='left', left_on='ticker', right_on='Added_Ticker')
print(merge1.shape)

merge2 = pd.merge(merge1, sp_removed, how='left', left_on='ticker', right_on='Removed_Ticker')
print(merge2.shape)

merge2.head(2)

merge2['date_filed_str'] = merge2['filedAt'].str[:10]
merge2['date_filed'] = pd.to_datetime(merge2['date_filed_str'], format='%Y-%m-%d')

#merge2.info()

merge2.head(2)

merge2['drop_row1'] = np.where( (merge2['Date_Added'].notna() & (merge2['date_filed'] < merge2['Date_Added'] ) ), 1, 0)

merge2['drop_row2'] = np.where( (merge2['Date_Removed'].notna() & (merge2['date_filed'] > merge2['Date_Removed']) ), 1, 0)

#merge2[merge2.drop_row1 == 1]
#merge2[merge2.drop_row2 == 1]

print(merge2.shape)

f2_all = merge2[merge2.drop_row1 != 1]
print(f2_all.shape)

f3_all = f2_all[f2_all.drop_row2 != 1]
print(f3_all.shape)

# apply lemmatizer to preprocessed text

lemma = WordNetLemmatizer()  

f3_all['unstemmed'] = f3_all.preproc_text.str.split()

f3_all['stem_text'] = f3_all.unstemmed.apply(lambda x: [lemma.lemmatize(y) for y in x])

f3_all['clean_text'] = f3_all.stem_text.apply(' '.join)

f3_all.head()

f3_all.columns

f4_all = f3_all[['date_filed', 'filedAt', 'formType', 'accessionNo', 'ticker', 'cik', 
                'companyName', 'companyNameLong', 'linkToTxt', 'formType',
                'sic', 'fiscalYearEnd', 'value', 'preproc_text', 'clean_text']]

with open('../data/preproc/preproc_text_all_years.pickle', 'wb') as to_write:
    pickle.dump(f4_all, to_write)
