==> code/08_more_text_cleaning.ipynb <==
''' more headings to remove '''

import pandas as pd
import numpy as np
import pickle
import datetime
import re

with open('../data/preproc/preproc_text_all_years.pickle', 'rb') as read_file:
    prepped_text = pickle.load(read_file)
    

prepped_text.columns

prepped_text.clean_text.str.contains('—').value_counts()

# replace hyphens with nothing
nodashes = lambda x: re.sub('—', '', x)

prepped_text.loc[:,'clean_text2'] = prepped_text.loc[:,'clean_text'].map(nodashes)

prepped_text.clean_text2.str.contains('—').value_counts()

headings =['item risk factors', 'items risk factors', 'item risk', 'other risk', 'at risk', 'risk measurement', 
          'risk identification and measurement', 'value at risk', 'value at risk VAR', 
          'risk factor', 'risk', 'risks', 'risk measurement and', 'other risks', 'risk factors and item',
          'and item risk factors', 'risk factors', 'business risk', 'business risks', 'item risk factors – other',
          'company risks', 'company risk factors', 'economic value of equity at risk', 
          'eographic risks', 'ervice risks', 'item a risk factors']

len(prepped_text[prepped_text['clean_text2'].str.lower().isin(headings)])

#prepped_text['clean_text3'] = prepped_text['clean_text2'].apply(lambda x: " ".join(x.split()))
 
f1_df = (prepped_text[~prepped_text['clean_text2'].str.lower().isin(headings)])
f1_df.shape

f1_df.drop('clean_text', axis=1, inplace=True)
f1_df.rename(columns={'clean_text2':'clean_text'}, inplace=True)

f1_df.head(2)

with open('../data/preproc/clean_text_all_years.pickle', 'wb') as to_write:
    pickle.dump(f1_df, to_write)
