==> code/09_NMF_topic_modeling.ipynb <==
"""
NMF topic modeling of each risk paragraph
2011-2020
"""

import pandas as pd
import numpy as np
import pickle
import datetime
import re
import nltk
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.decomposition import NMF, LatentDirichletAllocation

with open('../data/preproc/clean_text_all_years.pickle', 'rb') as read_file:
    prepped_text = pickle.load(read_file)

prepped_text.shape

prepped_text.columns

# add additional stopwords
stopwords = nltk.corpus.stopwords.words('english')
stopwords.extend(['risk', 'risks', 'factor', 'factors', 'item', 'quantitative', 'qualitative',
                 'company', 'business', 'firm', 'us', 'our', 
                'could', 'may', 'quarter', 'year'
                "january", "february", "march", "april", "may", "june", 
                "july", "august", "september", "october", "november", "december",
                "jan", "feb", "mar", "apr", "jun", "jul", "aug", "sep", "oct", "nov", "dec",
                "million", "thousand", "hundred", "billion"])

# Count Vectorizer with bi-grams
vectorizer = CountVectorizer(stop_words = stopwords, ngram_range=(1,2))
risk_words = vectorizer.fit_transform(prepped_text.clean_text)
risk_words.shape

nmf_model = NMF(20) 
risk_topic20 = nmf_model.fit_transform(risk_words)

words = vectorizer.get_feature_names()
t = nmf_model.components_.argsort(axis=1)[:,-1:-11:-1]
topic_words20 = [[words[e] for e in l] for l in t]
topic_words20

nmf_model = NMF(22) 
risk_topic22 = nmf_model.fit_transform(risk_words)

words = vectorizer.get_feature_names()
t = nmf_model.components_.argsort(axis=1)[:,-1:-11:-1]
topic_words22 = [[words[e] for e in l] for l in t]
topic_words22

nmf_model = NMF(24) 
risk_topic24 = nmf_model.fit_transform(risk_words)

words = vectorizer.get_feature_names()
t = nmf_model.components_.argsort(axis=1)[:,-1:-11:-1]
topic_words24 = [[words[e] for e in l] for l in t]
topic_words24

nmf_model = NMF(30) 
risk_topic30 = nmf_model.fit_transform(risk_words)

words = vectorizer.get_feature_names()
t = nmf_model.components_.argsort(axis=1)[:,-1:-11:-1]
topic_words30 = [[words[e] for e in l] for l in t]
topic_words30

# after review decide on 22 topics

prepped_text['risk_topic_nmf'] = risk_topic22.argmax(axis=1)

risks_info = prepped_text.reset_index()

components = pd.DataFrame(risk_topic22.round(5),
             index = prepped_text.clean_text,
             columns = ["C_0", "C_1","C_2", "C_3", "C_4", "C_5", "C_6", "C_7", "C_8", "C_9", 
                       "C_10", "C_11", "C_12", "C_13", "C_14", "C_15", "C_16", "C_17", "C_18",
                       "C_19", "C_20", "C_21"])

c_info = components.reset_index()

c_info.head(2)

risks_info.shape

c_info.shape

# add the topic components to the full df

all_df = pd.concat([risks_info, c_info], axis=1)
all_df.shape

all_df.columns

all_df_review = all_df[['date_filed', 'ticker', 'companyName', 'value', 'risk_topic_nmf']]
all_df_review.to_excel('review_22_nmf_topics.xlsx', index=False)

######
# CountVectorizer and NMF with 22 topics
topic_names = {0:"Fair Value", 1:"Accounting", 2:"Other", 3:"Interest Rate", 
               4:"Disclosure", 5:"Foreign Exchange ", 6:"Credit", 
               7:"Investment", 8:"Stock", 9:"Capital",
              10:"Liquidity1", 11:"Market", 12:"Derivatives", 13:"Internal Control", 
               14:"Liquidity2", 15:"Loans", 16:"Security", 17: "Product Development", 
               18:"Risk Management", 19:"Insurance", 20:"Commodity", 21: "Regulation"}



all_df["risk_topics"] = all_df.risk_topic_nmf.map(topic_names)
all_df.head(2)

with open('../data/topics/NMF_22_risk_topics_FINAL.pickle', 'wb') as to_write:
    pickle.dump(all_df, to_write)

#################

## Additional topic modeling with TF-IDF

#### TRY TFIDF and NMF
TFvectorizer = TfidfVectorizer(stop_words=stopwords, ngram_range=(1,2))
TFrisk_words = TFvectorizer.fit_transform(prepped_text.clean_text)
TFrisk_words.shape

nmf_model = NMF(22) 
TFrisk_topic22 = nmf_model.fit_transform(TFrisk_words)

words = vectorizer.get_feature_names()
t = nmf_model.components_.argsort(axis=1)[:,-1:-11:-1]
TFtopic_words22 = [[words[e] for e in l] for l in t]
TFtopic_words22

nmf_model = NMF(25) 
TFrisk_topic25 = nmf_model.fit_transform(TFrisk_words)

words = vectorizer.get_feature_names()
t = nmf_model.components_.argsort(axis=1)[:,-1:-11:-1]
TFtopic_words25 = [[words[e] for e in l] for l in t]
TFtopic_words25

nmf_model = NMF(12)  
TFrisk_topic12 = nmf_model.fit_transform(doc_word)

words = vectorizer.get_feature_names()
t = nmf_model.components_.argsort(axis=1)[:,-1:-11:-1]
TFtopic_words12 = [[words[e] for e in l] for l in t]
TFtopic_words12
