==> code/11_risks_over_time.ipynb <==
"""
group data to plot relative frequency of risks over time
2011-2020
"""

import pickle
import pandas as pd
import numpy as np
import datetime as dt

with open('../data/topics/NMF_22_risk_topics_FINAL.pickle', 'rb') as read_file:
    risk_topics = pickle.load(read_file)

risk_topics.head(2)

# year column
risk_topics['year'] = risk_topics['date_filed'].dt.year
risk_topics.head(2)

risk_topics.year.unique()

risk_topics.columns

year_agg =risk_topics[['year', 'risk_topic_nmf', 'risk_topics', 'value']].groupby(['year', 'risk_topic_nmf','risk_topics']).count().reset_index()

year_agg.head()

year_agg.tail()

year_totals = year_agg[['year', 'value']].groupby(['year']).agg({'value':np.sum}).reset_index()

year_totals.columns = ['year', 'year_total']
year_totals

by_year = pd.merge(year_agg, year_totals, how='left', on='year')
by_year.head()

by_year['freq'] = (by_year['value'] / by_year['year_total']) *100
by_year.head()

by_year.head(25)

by_year.to_excel('by_year_table.xlsx', index=False)

pivot_year = by_year.pivot_table(index=['risk_topics'], columns='year', values='freq', fill_value=0).reset_index()
pivot_year.head()

pivot_year.to_excel('pivot_year_table.xlsx', index=False)

pivot_year2 = by_year.pivot_table(index=['year'], columns='risk_topics', values='freq', fill_value=0).reset_index()
pivot_year2.head()

pivot_year2.to_excel('pivot_year_table2.xlsx', index=False)
