==> code/12_kmeans_clustering.ipynb <==
""" K-means clustering of 22 NMF topic components PARAGRAPHS"""

import pickle
import pandas as pd
import numpy as np
from sklearn.cluster import KMeans
import matplotlib.pyplot as plt
import seaborn as sns

with open('../data/topics/NMF_22_risk_topics_FINAL.pickle', 'rb') as read_file:
    risk_topics = pickle.load(read_file)

risk_topics.columns

X = np.array(risk_topics[['C_0', 'C_1', 'C_2', 'C_3', 'C_4', 'C_5', 'C_6', 'C_7', 'C_8', 'C_9',
       'C_10', 'C_11', 'C_12', 'C_13', 'C_14', 'C_15', 'C_16', 'C_17', 'C_18', 'C_19', 'C_20', 'C_21']])

# helper function that allows us to display data in 2 dimensions an highlights the clusters
def display_cluster(X,km=[],num_clusters=0):
    color = 'brgcmyk'
    alpha = 0.5
    s = 20
    if num_clusters == 0:
        plt.scatter(X[:,0],X[:,1],c = color[0],alpha = alpha,s = s)
    else:
        for i in range(num_clusters):
            plt.scatter(X[km.labels_==i,0],X[km.labels_==i,1],c = color[i],alpha = alpha,s=s)
            plt.scatter(km.cluster_centers_[i][0],km.cluster_centers_[i][1],c = color[i], marker = 'x', s = 100)

num_clusters = 11
km = KMeans(n_clusters=num_clusters,random_state=10,n_init=5) # n_init, number of times the K-mean algorithm will run
km.fit(X)

#display_cluster(X,km,num_clusters)

clusters = pd.DataFrame(data=km.labels_.tolist(), columns=['cluster'])

order_centroids = km.cluster_centers_.argsort()[:, ::-1] 
order_centroids[0]

risks_clusters = pd.concat([risk_topics, clusters], axis=1)

risks_clusters.shape

risks_clusters.columns

subset = risks_clusters[['value', 'risk_topic_nmf', 'risk_topics', 'cluster']]

subset[subset['cluster']==0].risk_topics.value_counts()

subset[subset['cluster']==1].risk_topics.value_counts()

subset[subset['cluster']==2].risk_topics.value_counts()

subset[subset['cluster']==3].risk_topics.value_counts()

subset[subset['cluster']==4].risk_topics.value_counts()

subset[subset['cluster']==5].risk_topics.value_counts()

subset[subset['cluster']==6].risk_topics.value_counts()

subset[subset['cluster']==7].risk_topics.value_counts()

subset[subset['cluster']==8].risk_topics.value_counts()

subset[subset['cluster']==9].risk_topics.value_counts()

subset[subset['cluster']==10].risk_topics.value_counts()

risks_clusters.head(2)

clusters_info = risks_clusters[['date_filed','ticker', 'companyName', 'sic',
      'value', 'risk_topic_nmf', 'C_0', 'C_1', 'C_2', 'C_3', 'C_4',
       'C_5', 'C_6', 'C_7', 'C_8', 'C_9', 'C_10', 'C_11', 'C_12', 'C_13',
       'C_14', 'C_15', 'C_16', 'C_17', 'C_18', 'C_19', 'C_20', 'C_21',
       'risk_topics', 'cluster']]

clusters_info.to_excel('risk_clusters_11.xlsx', index=False)
