==> code/13_agg_risk_topics_report.ipynb <==
""" 
topic modeling was done for each paragraph
weight each paragraph topic component by number of words in paragraph
calc weighted average of each topic for each report 
"""

import pandas as pd
import numpy as np
import pickle

with open('../data/topics/NMF_22_risk_topics_FINAL.pickle', 'rb') as read_file:
    risk_topics = pickle.load(read_file)

topics = risk_topics.reset_index()

topics.columns = ['level_0', 'index', 'date_filed', 'filedAt', 'formType', 'accessionNo',
       'ticker', 'cik', 'companyName', 'companyNameLong', 'linkToTxt',
       'formType2', 'sic', 'fiscalYearEnd', 'value', 'preproc_text',
       'clean_text', 'risk_topic_nmf', 'clean_text2', 'C_0', 'C_1', 'C_2',
       'C_3', 'C_4', 'C_5', 'C_6', 'C_7', 'C_8', 'C_9', 'C_10', 'C_11', 'C_12',
       'C_13', 'C_14', 'C_15', 'C_16', 'C_17', 'C_18', 'C_19', 'C_20', 'C_21',
       'risk_topics']

topics.drop(['level_0', 'index'], inplace=True, axis=1)

topics.head(2)

topics['word_count'] = topics['clean_text'].apply(lambda x: len(x))

topics.head(2)

topics['formType'].unique()

f1_topics = topics[topics['formType'].isin(['10-K'])]

weight_topics = f1_topics.copy()


#weight_topics.loc[:,'C_0':'C_21'] = topics.loc[:,'C_0':'C_21'] * topics['word_count']
#above command was too much for ram

weight_topics['C_0_N'] = weight_topics['C_0'] * topics['word_count']
weight_topics['C_1_N'] = weight_topics['C_1'] * topics['word_count']
weight_topics['C_2_N'] = weight_topics['C_2'] * topics['word_count']
weight_topics['C_3_N'] = weight_topics['C_3'] * topics['word_count']
weight_topics['C_4_N'] = weight_topics['C_4'] * topics['word_count']
weight_topics['C_5_N'] = weight_topics['C_5'] * topics['word_count']
weight_topics['C_6_N'] = weight_topics['C_6'] * topics['word_count']
weight_topics['C_7_N'] = weight_topics['C_7'] * topics['word_count']
weight_topics['C_8_N'] = weight_topics['C_8'] * topics['word_count']
weight_topics['C_9_N'] = weight_topics['C_9'] * topics['word_count']
weight_topics['C_10_N'] = weight_topics['C_10'] * topics['word_count']
weight_topics['C_11_N'] = weight_topics['C_11'] * topics['word_count']
weight_topics['C_12_N'] = weight_topics['C_12'] * topics['word_count']
weight_topics['C_13_N'] = weight_topics['C_13'] * topics['word_count']
weight_topics['C_14_N'] = weight_topics['C_14'] * topics['word_count']
weight_topics['C_15_N'] = weight_topics['C_15'] * topics['word_count']
weight_topics['C_16_N'] = weight_topics['C_16'] * topics['word_count']
weight_topics['C_17_N'] = weight_topics['C_17'] * topics['word_count']
weight_topics['C_18_N'] = weight_topics['C_18'] * topics['word_count']
weight_topics['C_19_N'] = weight_topics['C_19'] * topics['word_count']
weight_topics['C_20_N'] = weight_topics['C_20'] * topics['word_count']
weight_topics['C_21_N'] = weight_topics['C_21'] * topics['word_count']


# year column
weight_topics['year'] = weight_topics['date_filed'].dt.year

weight_topics.columns

# average the weighted topic componenets by year, date_filed, accessionNo, ticker, companyName 
weight_topics2 = weight_topics[['year', 'date_filed', 'accessionNo', 'formType', 'ticker', 
                                'companyName','sic',
                              'C_0_N', 'C_1_N', 'C_2_N', 'C_3_N',
                               'C_4_N', 'C_5_N', 'C_6_N', 'C_7_N', 'C_8_N', 'C_9_N', 'C_10_N',
                               'C_11_N', 'C_12_N', 'C_13_N', 'C_14_N', 'C_15_N', 'C_16_N', 'C_17_N',
                               'C_18_N', 'C_19_N', 'C_20_N', 'C_21_N']]

weight_topics2.head(2)


report_topics = weight_topics2.groupby(['year', 'date_filed', 'accessionNo', 'formType', 'sic', 'ticker', 'companyName']).mean().reset_index()


report_topics.head(2)

report_topics.ticker.value_counts()

#report_topics.to_excel('review_topics_report_level.xlsx', index=False)

report_topics['ticker_year'] = report_topics['ticker'] + '_' + report_topics['year'].astype(str)

report_topics.head()

with open('../data/topics/NMF_report_risk_topics.pickle', 'wb') as to_write:
    pickle.dump(report_topics, to_write)
