==> code/14_kmeans_clustering_reports.ipynb <==
""" 
K-means clustering of 22 NMF topic components for each REPORT 
"""

import pickle
import pandas as pd
import numpy as np
from sklearn.cluster import KMeans
import matplotlib.pyplot as plt
import seaborn as sns

with open('../data/topics/NMF_report_risk_topics.pickle', 'rb') as read_file:
    risk_topics = pickle.load(read_file)

risk_topics.columns

X = np.array(risk_topics[['C_0_N', 'C_1_N', 'C_2_N', 'C_3_N', 'C_4_N', 'C_5_N',
       'C_6_N', 'C_7_N', 'C_8_N', 'C_9_N', 'C_10_N', 'C_11_N', 'C_12_N', 'C_13_N', 
       'C_14_N', 'C_15_N', 'C_16_N', 'C_17_N', 'C_18_N', 'C_19_N', 'C_20_N', 'C_21_N']])

# helper function that allows us to display data in 2 dimensions an highlights the clusters
def display_cluster(X,km=[],num_clusters=0):
    color = 'brgcmyk'
    alpha = 0.5
    s = 20
    if num_clusters == 0:
        plt.scatter(X[:,0],X[:,1],c = color[0],alpha = alpha,s = s)
    else:
        for i in range(num_clusters):
            plt.scatter(X[km.labels_==i,0],X[km.labels_==i,1],c = color[i],alpha = alpha,s=s)
            plt.scatter(km.cluster_centers_[i][0],km.cluster_centers_[i][1],c = color[i], marker = 'x', s = 100)

num_clusters = 11
km = KMeans(n_clusters=num_clusters,random_state=10,n_init=5) # n_init, number of times the K-mean algorithm will run
km.fit(X)

#display_cluster(X,km,num_clusters)

clusters = pd.DataFrame(data=km.labels_.tolist(), columns=['cluster'])

order_centroids = km.cluster_centers_.argsort()[:, ::-1] 
order_centroids[0]

risks_clusters = pd.concat([risk_topics, clusters], axis=1)

risks_clusters.columns

risks_clusters.head()

risks_clusters.to_excel('report_clusters11.xlsx', index=False)

#######
num_clusters = 10
km = KMeans(n_clusters=num_clusters,random_state=10,n_init=10) # n_init, number of times the K-mean algorithm will run
km.fit(X)

clusters = pd.DataFrame(data=km.labels_.tolist(), columns=['cluster'])
risks_clusters = pd.concat([risk_topics, clusters], axis=1)
risks_clusters.to_excel('report_clusters10.xlsx', index=False)

km.cluster_centers_

#######
num_clusters = 12
km = KMeans(n_clusters=num_clusters,random_state=10,n_init=10) # n_init, number of times the K-mean algorithm will run
km.fit(X)
clusters = pd.DataFrame(data=km.labels_.tolist(), columns=['cluster'])
risks_clusters = pd.concat([risk_topics, clusters], axis=1)
risks_clusters.to_excel('report_clusters12.xlsx', index=False)

#######
num_clusters = 9
km = KMeans(n_clusters=num_clusters,random_state=10,n_init=10) # n_init, number of times the K-mean algorithm will run
km.fit(X)
clusters = pd.DataFrame(data=km.labels_.tolist(), columns=['cluster'])
risks_clusters = pd.concat([risk_topics, clusters], axis=1)
risks_clusters.to_excel('report_clusters9.xlsx', index=False)

#######
num_clusters = 8
km = KMeans(n_clusters=num_clusters,random_state=10,n_init=10) # n_init, number of times the K-mean algorithm will run
km.fit(X)
clusters = pd.DataFrame(data=km.labels_.tolist(), columns=['cluster'])
risks_clusters = pd.concat([risk_topics, clusters], axis=1)
risks_clusters.to_excel('report_clusters8.xlsx', index=False)

#######
num_clusters = 7
km = KMeans(n_clusters=num_clusters,random_state=10,n_init=10) # n_init, number of times the K-mean algorithm will run
km.fit(X)
clusters = pd.DataFrame(data=km.labels_.tolist(), columns=['cluster'])
risks_clusters = pd.concat([risk_topics, clusters], axis=1)
risks_clusters.to_excel('report_clusters7.xlsx', index=False)

#km.cluster_centers_

#######
num_clusters = 6
km = KMeans(n_clusters=num_clusters,random_state=10,n_init=10) # n_init, number of times the K-mean algorithm will run
km.fit(X)
clusters = pd.DataFrame(data=km.labels_.tolist(), columns=['cluster'])
risks_clusters = pd.concat([risk_topics, clusters], axis=1)
risks_clusters.to_excel('report_clusters6.xlsx', index=False)

# 6 is too few - the banks and insurances were dipersed
